Modularize DrawerNavigator
